requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
pypdf>=4.0.0
python-docx>=1.1.0
plotly>=5.0.0
openpyxl>=3.1.0
python-pptx>=0.6.23
